Change name of the Code Comittee to Software Comittee
Rationale: We've been using the name Software Committee for a while (at least in some context), and my feeling is the word "software" is much less ambiguous than "code", thus making it more clear what the TC is for.

Note: According to the charter, this modification needs two thirds of the GB, so this is really a proposal to be discussed in the next GB meeting. Comments to this pull request are welcome in advance to that discussion.