Streamline Quickstart
Based on conversations in the DEI working group and Community Calls, we wanted to streamline the  Quickstart and identify three steps a newcomer can take to get acquainted with CHAOSS. These are the three we identified. I'd like to move some of the info that was previously in this Quickstart into maybe a new doc called "Next Steps: Participating in CHAOSS" or something like that, as we shift to more of a learning path for our contributor funnel. 

Signed-off-by: Elizabeth Barron <elizabeth@naramore.net>